+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.3.8] - in progress
+
+- New module `cloudly.util.http_client`.
+- New module `cloudly.util.multidownloader`.
+
+
 ## [0.3.7] - 2025-09-13
 
 - `cloudly.upathlib.Upath` gets two new methods `read_csv` and `write_csv`.